+++ packages/desktop/views/dashboard/drawers/contact-book/views/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ContactList } from './ContactList.svelte'
